pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "config-migrator"
version = "1.0.0"
description = "Configuration migration tool for YAML files with NSTF precedence"
authors = [{name = "Config Migrator Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ruamel.yaml>=0.18.0",
    "click>=8.0.0",
    "rich>=13.0.0"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[project.scripts]
config-migrator = "config_migrator.__main__:main"

[tool.mypy]
python_version = "3.8"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
